Changed project structure. Added CMakeLists.txt file for my_client example
